requirements.txt
streamlit==0.71.0
joblib==0.14.1
pandas==1.1.4
category_encoders==2.2.2
memoized-property==1.0.3
termcolor==1.1.0
numpy==1.18.5
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
matplotlib==3.3.3
mlflow==1.8.0
numpy==1.18.5
xgboost==1.2.1
scipy==1.5.4
google-cloud-storage==1.26.0
mlflow==1.8.0
gcsfs==0.6.0
psutil==5.7.0
imblearn
Pillow==8.0.1
